fix: Windows build — rename resvg.lib to libresvg.lib and cast C enum rawValue to UInt32
SPM adds a "lib" prefix when linking, so it searches for libresvg.lib
instead of resvg.lib. Also, MSVC imports C enums as Int32 (signed)
while Swift enums are declared as UInt32 — explicit UInt32() cast
fixes the type mismatch (noop on macOS where it's already UInt32).

+++ Sources/Resvg/Gradient.swift
@@ -49,7 +49,7 @@ public struct LinearGradient: @unchecked Sendable {
 
     /// The spread method.
     public var spreadMethod: SpreadMethod {
-        SpreadMethod(rawValue: resvg_linear_gradient_spread_method(ptr).rawValue) ?? .pad
+        SpreadMethod(rawValue: UInt32(resvg_linear_gradient_spread_method(ptr).rawValue)) ?? .pad
     }
 
     /// The number of gradient stops.
@@ -135,7 +135,7 @@ public struct RadialGradient: @unchecked Sendable {
 
     /// The spread method.
     public var spreadMethod: SpreadMethod {
-        SpreadMethod(rawValue: resvg_radial_gradient_spread_method(ptr).rawValue) ?? .pad
+        SpreadMethod(rawValue: UInt32(resvg_radial_gradient_spread_method(ptr).rawValue)) ?? .pad
     }
 
     /// The number of gradient stops.
